+++ helpers/deprecated/pick.js
@@ -1,13 +1,23 @@
 'use strict';
 
-const _ = require('lodash');
-
 const factory = () => {
     /**
      * @deprecate
      */
-    return function() {
-        return _.pick.apply(null, arguments);
+    return function(obj, ...args) {
+        return args.reduce(
+            (
+                pickedElements,
+                element
+            ) => {
+                if (obj && obj.hasOwnProperty(element)) {
+                    pickedElements[element] = obj[element];
+                }
+
+                return pickedElements;
+            },
+            {}
+        );
     };
 };