+++ helpers/resourceHints.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const _ = require('lodash');
 const getFonts = require('./lib/fonts');
 
 const fontResources = {
@@ -10,24 +9,27 @@ const fontResources = {
     ],
 };
 
+function format(host) {
+    return `<link rel="dns-prefetch preconnect" href="${host}" crossorigin>`;
+}
+
 const factory = globals => {
     return function() {
-        function format(host) {
-            return `<link rel="dns-prefetch preconnect" href="${host}" crossorigin>`;
-        }
-
         var hosts = [];
 
         // Add cdn
-        const siteSettings = globals.getSiteSettings();
-        const cdnUrl = siteSettings['cdn_url'] || '';
-        if (cdnUrl != '') {
+        const cdnUrl = globals.getSiteSettings()['cdn_url'];
+        if (cdnUrl) {
             hosts.push(cdnUrl);
         }
 
-        // Add font providers
-        const fontProviders = _.keys(getFonts('providerLists', globals.getThemeSettings(), globals.handlebars));
-        _.each(fontProviders, function(provider) {
+        const fonts = getFonts(
+            'providerLists',
+            globals.getThemeSettings(),
+            globals.handlebars
+        );
+
+        Object.keys(fonts).forEach((provider) => {
             if (typeof fontResources[provider] !== 'undefined') {
                 hosts = hosts.concat(fontResources[provider]);
             }